Part 1

+++ example/binomial_confidence_limits.cpp
@@ -57,7 +57,7 @@ void confidence_limits_on_frequency(unsigned trials, unsigned successes)
    //
    // Now print out the data for the table rows.
    //
-   for(unsigned i = 0; i < sizeof(alpha)/sizeof(alpha[0]); ++i)
+   for(size_t i = 0; i < sizeof(alpha)/sizeof(alpha[0]); ++i)
    {
       // Confidence value:
       cout << fixed << setprecision(3) << setw(10) << right << 100 * (1-alpha[i]);
